Added WheelController

+++ Backend/API/API/Models/Input/VehicleCreateModel.cs
@@ -17,6 +17,6 @@ namespace API.Models
 
         public string LocationId { get; set; }
 
-        public int Price { get; set; }
+        public float Price { get; set; }
     }
 }
